GenericTest15: B1Testプロジェクト追加

+++ GenericTest/GenericTest15/ServiceApi/Requests/A1/A1Request.cs
@@ -2,5 +2,5 @@
 
 public class A1Request : RequestBase
 {
-    public required int A1Value { get; init; }
+    public required decimal A1Value { get; init; }
 }
